Added "version" command Replaced unsigned int by size_t

+++ src/exe/ModelOptimize.h
@@ -29,7 +29,7 @@ public:
 	int optimizePartitioningScheme(PartitioningScheme * scheme, int index=0, int limit=1);
 	int optimizePartitionElement(PartitionElement * scheme, int index=0, int limit=1);
 private:
-	void optimizeModel(PartitionElement * element, unsigned int modelIndex, int limit);
+	void optimizeModel(PartitionElement * element, size_t modelIndex, int limit);
 	void setModelParameters(Model * _model, pllInstance * _tree,
 			partitionList * _partitions, pllAlignmentData * _alignData, int index, bool setAlphaFreqs);
 };
